split local storage module
