+++ frontend/src/browser/store/local/create-function.ts
@@ -1,11 +1,19 @@
-import { IFunction } from "#types";
+import type { IFunction } from "#types";
+import { runMigrations } from "./migrations";
 
 let isMigrated = false;
 
-export function createLocalStorageFunction(func: IFunction) {
-  function storageFunction(
+export function createLocalStorageFunction(func: IFunction): typeof func {
+  async function storageFunction(
     ...args: Parameters<typeof func>
-  ): ReturnType<typeof func> {}
+  ): Promise<ReturnType<typeof func>> {
+    if (!isMigrated) {
+      await runMigrations();
+      isMigrated = true;
+    }
+
+    return await func(...args);
+  }
 
   return storageFunction;
 }
