+++ frontend/src/browser/store/local/index.ts
@@ -1,6 +1,1 @@
-export {
-  isLocalStorageAvailable,
-  getLocalStoreItem,
-  setLocalStoreItem
-} from "./local-storage";
-export type { ILocalStoreKey } from "./local-storage";
+export type { ILocalStoreKey } from "./types";
