+++ frontend/src/entities/task/lib/get.ts
@@ -6,7 +6,6 @@ import { createValidator } from "#lib/json/schema";
 import { getLocalStoreItem } from "#store/local";
 import { IPlace, getAllPlaces } from "#entities/place";
 import type { ITask, ITasksStats, ITaskStore } from "../types";
-import { migrateTasks } from "./migrate";
 import { toTasks } from "./to-tasks";
 
 interface IOptions {
@@ -17,8 +16,6 @@ interface IOptions {
   placeID?: IPlace["id"];
 }
 
-let isMigrated = false;
-
 const defaultOptions = {
   includeDeleted: false,
 } as const satisfies IOptions;
@@ -26,14 +23,6 @@ const defaultOptions = {
 const validateTask = createValidator<ITaskStore>("/entities/task/entity");
 
 export async function getTask(taskID: ITask["id"]): Promise<ITask> {
-  if (!isMigrated) {
-    try {
-      await migrateTasks();
-    } finally {
-      isMigrated = true;
-    }
-  }
-
   const storedTasks = await getAllTasks();
   const storeTask = storedTasks.find(({ id }) => id === taskID);
 
@@ -117,14 +106,6 @@ export async function getTasks(
   options: IOptions = defaultOptions,
 ): Promise<IPaginatedCollection<ITask>> {
   logDebug(`Getting tasks...`);
-
-  if (!isMigrated) {
-    try {
-      await migrateTasks();
-    } finally {
-      isMigrated = true;
-    }
-  }
 
   const {
     includeDeleted,
